monitoring/requirements.txt
requests==2.31.0
PyGithub==2.1.1
psycopg2-binary==2.9.9
